pyproject.toml
[project]
name = "crypto-predictor-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "candles",
    "confluent-kafka>=2.8.2",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
    "quixstreams>=3.13.1",
    "technical_indicators",
    "trades",
    "websocket-client>=1.8.0",
]

[tool.uv.workspace]
members = ["services/trades", "services/candles", "services/technical_indicators", "services/predictor"]

[tool.uv.sources]
trades = { workspace = true }
candles = { workspace = true }
technical-indicators = { workspace = true }

[tool.ruff.lint]
select = [
  "F", # pyflakes rules
  "C", # mccabe rules
  "E", # pycodestyle error rules
  "W", # pycodestyle warning rules
  "B", # flake8-bugbear rules
  "I", # isort rules
]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pycodestyle]
# Ignore overlong task comments
ignore-overlong-task-comments = true
